separate config from NotificationApplication for loose coupling unit test purpose
