+++ notification-service/src/main/java/com/piggymetrics/notification/config/ResourceServerConfig.java
@@ -1,0 +1,34 @@
+package com.piggymetrics.notification.config;
+
+import feign.RequestInterceptor;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.cloud.security.oauth2.client.feign.OAuth2FeignRequestInterceptor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.oauth2.client.DefaultOAuth2ClientContext;
+import org.springframework.security.oauth2.client.OAuth2RestTemplate;
+import org.springframework.security.oauth2.client.token.grant.client.ClientCredentialsResourceDetails;
+import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer;
+import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfigurerAdapter;
+
+/**
+ * @author cdov
+ */
+@Configuration
+@EnableResourceServer
+public class ResourceServerConfig extends ResourceServerConfigurerAdapter {
+    @Bean
+    @ConfigurationProperties(prefix = "security.oauth2.client")
+    public ClientCredentialsResourceDetails clientCredentialsResourceDetails() {
+        return new ClientCredentialsResourceDetails();
+    }
+    @Bean
+    public RequestInterceptor oauth2FeignRequestInterceptor(){
+        return new OAuth2FeignRequestInterceptor(new DefaultOAuth2ClientContext(), clientCredentialsResourceDetails());
+    }
+
+    @Bean
+    public OAuth2RestTemplate clientCredentialsRestTemplate() {
+        return new OAuth2RestTemplate(clientCredentialsResourceDetails());
+    }
+}
